Add files via upload
Version 1 of ResNet18 modified to take an additional tabular input in addition to an image. The architecture is modified to include a separate MLP branch to extract information from the tabular data in addition to the standard architecture consisting of basic blocks that extracts from the x-ray image. Once features are extracted from both modalities of data, the features are concatenated together and fed through a final MLP and output to classify the presence of 14 diseases (non-mutually-exclusive).